refactor(orc8r): Split state to protected and southbound
