+++ orc8r/cloud/go/services/state/servicers/protected/indexer_mgr_servicer.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package servicers
+package protected
 
 import (
 	"context"
