+++ orc8r/cloud/go/services/state/servicers/protected/indexer_mgr_servicer_test.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package servicers_test
+package protected_test
 
 import (
 	"context"
